% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_external_data.R
\name{get_external_data}
\alias{get_external_data}
\title{Access external test data for the otsuSeg package}
\usage{
get_external_data(data_type)
}
\arguments{
\item{data_type}{Character string specifying the type of data to retrieve.
One of "NBRpre", "NBRpost", or "shapefile_reference".}
}
\value{
Returns the file path of the requested data.
}
\description{
This function helps users easily access the test data provided with the otsuSeg package.
}
\examples{
# Get file path
get_external_data("NBRpre")

# Load raster manually
raster_data <- raster::raster(get_external_data("NBRpre"))
plot(raster_data)

# Load shapefile manually
shape_data <- sf::st_read(get_external_data("shapefile_reference"))
plot(shape_data)
}
