% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_external_data.R
\name{get_external_data}
\alias{get_external_data}
\title{Download and load example data for OtsuSeg}
\usage{
get_external_data(filename = NULL, path = tempdir(), load = FALSE)
}
\arguments{
\item{filename}{Optional. The name of the file to return (after extraction).}

\item{path}{Local directory to download and extract files. Default is a temp folder.}

\item{load}{Logical. If TRUE, returns a loaded raster object (only for .tif). Default is FALSE.}
}
\value{
File path or loaded object if `load = TRUE`.
}
\description{
Downloads a ZIP file with example rasters from GitHub releases and optionally loads a specific file.
}
